specify c++ standard

+++ tutorial/tutorial.cc
@@ -1,5 +1,4 @@
 #include <cmath>
-#include <cstdlib>
 #include <iostream>
 #include <string>
 
@@ -16,7 +15,7 @@ int main(int argc, char* argv[])
   }
 
   // convert input to double
-  const double inputValue = atof(argv[1]);
+  const double inputValue = std::stof(argv[1]);
 
   // calculate square root
   const double outputValue = sqrt(inputValue);
